[turbopack] Move require.context types into well_known/require_context.rs
